Create clipboard_sort
This is a program that I wrote to resolve a few different challenges that I was seeing across hacking-practice websites. The challenges usually involve sorting either number or words and submitting them sorted within a 5 second time frame. Rather than writing a program that accessed the website, authenticated, navigated to the challenge, located the proper tags, sorted the proper information, and submitted the proper information, I decided it would be easier to copy the information with my mouse+keyboard, run my program, paste into the form, and submit manually. 

This program will alphabetize the words in your clipboard and separate them with comma's (and it makes everything lowercase) then add them back to your clipboard.