Psuedo-reverse shell for HTB's Baby Website Rick
This is a psuedo-reverse shell attacking the deserialization vulnerability in HackTheBox's 'Baby Website Rick.' It makes use of the Python (2.7) pickling modules, requests, and some <i>bad</i> parsing to return output to users. It's setup -by default- to proxy your requests through BurpSuite (for troubleshooting), so comment out the line, and change the URL to use it properly.